refactor: fix plugin internal style errors
