+++ Dalamud/Interface/Internal/Windows/PluginInstaller/PluginChangelogEntry.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 
 using Dalamud.Plugin.Internal;
+using Dalamud.Plugin.Internal.Types;
 using Dalamud.Utility;
 
 namespace Dalamud.Interface.Internal.Windows.PluginInstaller
