+++ Dalamud/Plugin/Internal/Types/PluginUpdateStatus.cs
@@ -1,30 +1,29 @@
 using System;
 
-namespace Dalamud.Plugin.Internal.Types
+namespace Dalamud.Plugin.Internal.Types;
+
+/// <summary>
+/// Plugin update status.
+/// </summary>
+internal class PluginUpdateStatus
 {
     /// <summary>
-    /// Plugin update status.
+    /// Gets the plugin internal name.
     /// </summary>
-    internal class PluginUpdateStatus
-    {
-        /// <summary>
-        /// Gets or sets the plugin internal name.
-        /// </summary>
-        public string InternalName { get; set; }
+    public string InternalName { get; init; } = null!;
 
-        /// <summary>
-        /// Gets or sets the plugin name.
-        /// </summary>
-        public string Name { get; set; }
+    /// <summary>
+    /// Gets the plugin name.
+    /// </summary>
+    public string Name { get; init; } = null!;
 
-        /// <summary>
-        /// Gets or sets the plugin version.
-        /// </summary>
-        public Version Version { get; set; }
+    /// <summary>
+    /// Gets the plugin version.
+    /// </summary>
+    public Version Version { get; init; } = null!;
 
-        /// <summary>
-        /// Gets or sets a value indicating whether the plugin was updated.
-        /// </summary>
-        public bool WasUpdated { get; set; }
-    }
+    /// <summary>
+    /// Gets or sets a value indicating whether the plugin was updated.
+    /// </summary>
+    public bool WasUpdated { get; set; }
 }
